Use DataFlowValue instead of Descriptor for equality of ESDataFlowValue
Equality of those values is used for keys in maps with data-flow info.
Those maps sometimes should be intersected (e.g. in boolean expressions
with '||').
E.g. for expression: 'this.x != null || other.x != null' we have to
intersect 2 maps as follows:
{ this.x : NON_NULL } intersect { other.x : NON_NULL}.

Using 'ValueDescriptor' for equality is obviously wrong, because then
'this.x' and 'other.x' will be treated as equal, and aforementioned
intersection will be non-empty, leading to unsound smartcasts *even
in the absence of contracted calls*.

This commit adds proper overload of 'equals' to 'ESDataFlowValue' that
will compare them based on underlying 'DataFlowValue' (which already
distinguishes those values)

NB. This unsoundness was introduced only since version 1.3 (because
in 1.2.X contracts were turned off), so fixing it isn't considered
to be breaking change

^KT-27260 Fixed